# REST_APIs_Design/openapi.yaml
openapi: 3.0.1
info:
   title: Film Manager
   description: REST APIs for the interaction with the Film Manager service.
   termsOfService: 'http://swagger.io/terms/'
   license:
      name: Apache 2.0
      url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
   version: 1.0.0
...
servers:
   - url: 'https://localhost/v1'
   - url: 'http://localhost/v1'
...
paths:
   '/api/films':
      post:
         tags:
            - films
         security:
            - cookieAuth: []
         summary: Create a new film
         description: A new film is created by the authenticated user.
         operationId: createFilm
         requestBody:
            description: Representation of the film to be created
            content:
               application/json:
                  schema:
                     $ref: '#/components/schemas/Film'
            required: true
         responses:
            '201':
               description: Created
               content:
                  application/json:
                     schema:
                        $ref: '#/components/schemas/Film'
            '400':
               description: Bad Request
               content: {}
            '401':
               description: Unauthorized
               content: {}
            '500':
               description: Internal Server Error
               content: {}
   '/api/films/private':
      get:
         tags:
            - films/private
         security:
            - cookieAuth: []
         summary: Retrieve the private films of the logged-in user
         description: >-
            The private films of the logged-in user are retrieved. A pagination mechanism is implemented to limit the size of messages.
         operationId: getPrivateFilms
         parameters:
            - in: query
              name: pageNo
              required: false
              schema:
                 type: integer
              description: 'The id of the requested page (if absent, the first page is returned)'
         responses:
            '200':
               description: Successful operation
               content:
                  application/json:
                     schema:
                        type: object
                        properties:
                           totalPages:
                              type: integer
                           currentPage:
                              type: integer
                           totalItems:
                              type: integer
                           films:
                              type: array
                              items:
                                 $ref: '#/components/schemas/Film'
                           next:
                              type: string
                              format: uri
            '401':
               description: Unauthorized
               content: {}
            '500':
               description: Internal Server Error
               content: {}
   '/api/films/private/{filmId}':
      get:
         tags:
            - 'films/private/{filmId}'
         security:
            - cookieAuth: []
         summary: Retrieve a private film
         description: The private film with ID filmId is retrieved. This operation can be performed on the film if the user who performs the operation is the film's owner.
         operationId: getPrivateFilm
         parameters:
            - name: filmId
              in: path
              description: ID of the film to retrieve
              required: true
              schema:
                 type: integer
                 format: int64
         responses:
            '200':
               description: Successful operation
               content:
                  application/json:
                     schema:
                        $ref: '#/components/schemas/Film'
            '401':
               description: Unauthorized
               content: {}
            '403':
               description: Forbidden
               content: {}
            '404':
               description: Not Found
               content: {}
            '500':
               description: Internal Server Error
               content: {}
      put:
         tags:
            - 'films/private/{filmId}'
         security:
            - cookieAuth: []
         summary: Update a private film
         description: The private film with ID filmId is updated. This operation does not allow changing its visibility. This operation can be performed only by the owner.
         operationId: updatePrivateFilm
         parameters:
            - name: filmId
              in: path
              description: ID of the film to update
              required: true
              schema:
                 type: integer
                 format: int64
         requestBody:
            description: The updated film object that needs to replace the old object
            content:
               application/json:
                  schema:
                     $ref: '#/components/schemas/Film'
            required: true
         responses:
            '204':
               description: No Content
               content: {}
            '400':
               description: Bad Request
               content: {}
            '401':
               description: Unauthorized
               content: {}
            '403':
               description: Forbidden
               content: {}
            '409':
               description: Conflict (attempt to make unallowed change, e.g., visibility)
               content: {}
            '404':
               description: Not Found
               content: {}
            '500':
               description: Internal Server Error
               content: {}
      delete:
         tags:
            - 'films/private/{filmId}'
         security:
            - cookieAuth: []
         summary: Delete a private film
         description: The private film with ID filmId is deleted. This operation can only be performed by the owner.
         operationId: deletePrivateFilm
         parameters:
            - name: filmId
              in: path
              description: ID of the film to delete
              required: true
              schema:
                 type: integer
                 format: int64
         responses:
            '204':
               description: Successful operation - No Content
               content: {}
            '401':
               description: Unauthorized
               content: {}
            '403':
               description: Forbidden
               content: {}
            '404':
               description: Not Found
               content: {}
            '500':
               description: Internal Server Error
               content: {}
   '/api/films/public':
      get:
         tags:
            - films/public
         summary: Retrieve the public films
         description: >-
            The public films (i.e., the films that are visible for all the users of the service) are retrieved. This operation does not require authentication. A pagination mechanism is implemented to limit the size of messages.
         operationId: getPublicFilms
         parameters:
            - in: query
              name: pageNo
              required: false
              schema:
                 type: integer
              description: 'The id of the requested page (if absent, the first page is returned)'
         responses:
            '200':
               description: Successful operation
               content:
                  application/json:
                     schema:
                        type: object
                        properties:
                           totalPages:
                              type: integer
                           currentPage:
                              type: integer
                           totalItems:
                              type: integer
                           films:
                              type: array
                              items:
                                 $ref: '#/components/schemas/Film'
                           next:
                              type: string
                              format: uri
            '500':
               description: Internal Server Error
               content: {}
   '/api/films/public/invited':
      get:
         tags:
            - films/public/invited
         security:
            - cookieAuth: []
         summary: Retrieve the public films that the logged-in user has been invited to review
         description: >-
            The public films that the logged-in user has been invited to review are retrieved. A pagination mechanism is implemented to limit the size of messages.
         operationId: getInvitedFilms
         parameters:
            - in: query
              name: pageNo
              required: false
              schema:
                 type: integer
              description: 'The id of the requested page (if absent, the first page is returned)'
         responses:
            '200':
               description: Successful operation
               content:
                  application/json:
                     schema:
                        type: object
                        properties:
                           totalPages:
                              type: integer
                           currentPage:
                              type: integer
                           totalItems:
                              type: integer
                           films:
                              type: array
                              items:
                                 $ref: '#/components/schemas/Review'
                           next:
                              type: string
                              format: uri
            '401':
               description: Unauthorized
               content: {}
            '500':
               description: Internal Server Error
               content: {}
   '/api/films/public/{filmId}':
      get:
         tags:
            - films/public/{filmId}
         summary: Retrieve a public film
         description: >-
            The public film with ID filmId is retrieved. This operation does not require authentication.
         operationId: getPublicFilm
         parameters:
            - name: filmId
              in: path
              description: ID of the film to retrieve
              required: true
              schema:
                 type: integer
                 format: int64
         responses:
            '200':
               description: Successful operation
               content:
                  application/json:
                     schema:
                        $ref: '#/components/schemas/Film'
            '404':
               description: Not Found
               content: {}
            '500':
               description: Internal Server Error
               content: {}
      put:
         tags:
            - 'films/public/{filmId}'
         security:
            - cookieAuth: []
         summary: Update a public film
         description: The public film with ID filmId is updated. This operation does not allow changing its visibility. This operation can be performed only by the owner.
         operationId: updatePublicFilm
         parameters:
            - name: filmId
              in: path
              description: ID of the film to update
              required: true
              schema:
                 type: integer
                 format: int64
         requestBody:
            description: The updated film object that needs to replace the old object
            content:
               application/json:
                  schema:
                     $ref: '#/components/schemas/Film'
            required: true
         responses:
            '204':
               description: No Content
               content: {}
            '400':
               description: Bad Request
               content: {}
            '401':
               description: Unauthorized
               content: {}
            '403':
               description: Forbidden
               content: {}
            '409':
               description: Conflict (attempt to make unallowed change, e.g., visibility)
               content: {}
            '404':
               description: Not Found
               content: {}
            '500':
               description: Internal Server Error
               content: {}
      delete:
         tags:
            - 'films/public/{filmId}'
         security:
            - cookieAuth: []
         summary: Delete a public film
         description: The public film with ID filmId is deleted. This operation can only be performed by the owner.
         operationId: deletePublicFilm
         parameters:
            - name: filmId
              in: path
              description: ID of the film to delete
              required: true
              schema:
                 type: integer
                 format: int64
         responses:
            '204':
               description: Successful operation - No Content
               content: {}
            '401':
               description: Unauthorized
               content: {}
            '403':
               description: Forbidden
               content: {}
            '404':
               description: Not Found
               content: {}
            '500':
               description: Internal Server Error
               content: {}
...
components:
   schemas:
...
      Film:
         title: Film
         description: JSON Schema for Film data structures
         type: object
         additionalProperties: false
         properties:
           "$schema":
             type: string
             description: path to the JSON Schema to use for the validation of the data structure
           id:
             type: integer
             description: unique identifier for the film
           title:
             type: string
             description: textual title of the film
           owner:
             type: integer
             description: unique identifier of the film owner
           private:
             type: boolean
             description: true if only the film owner can have access to it, false otherwise
             default: true
           watchDate:
             type: string
             format: date
             description: the date when the film was watched by the owner
           rating:
             type: integer
             description: the rating the owner has given to the film
             minimum: 1
             maximum: 10
           favorite:
             type: boolean
             description: 'true if the film is among the favourite ones of the owner, false
               otherwise '
             default: false
         required:
         - id
         - title
         - owner
         - private
      Review:
         title: Review
         description: JSON Schema for Review data structures
         type: object
         additionalProperties: false
         properties:
           "$schema":
             type: string
             description: path to the JSON Schema to use for the validation of the data structure
           id:
             type: integer
             description: unique identifier for the review
           filmId:
             type: integer
             description: unique identifier of the film for which this review invitation has
               been issued
           reviewerId:
             type: integer
             description: unique identifier of the the user who has received this review invitation
           reviewersId:
             type: array
             description: the list of identifier of the the users which have received this
               review invitation
             items:
               type: integer
             minItems: 2
           completed:
             type: boolean
             description: true if the review has been completed, false otherwise
             default: false
           reviewDate:
             type: string
             format: date
             description: the date when the review has been completed by the invited user
           rating:
             type: integer
             description: the rating the reviewer has given to the film
             minimum: 1
             maximum: 10
           coop:
             type: boolean
             description: whether the review is a co-op one or not
             default: false
           review:
             type: string
             description: a textual description of the review
             maxLength: 1000
         required:
         - filmId
...
   securitySchemes:
      cookieAuth:
         description: JWT to identify the authenticated user.
         type: apiKey
         in: cookie
         name: connect.sid